--- routes/birds.js
@@ -1,6 +1,0 @@
-import {Router} from 'express';
-const router = Router();
-
-
-
-export default router;
