+++ routes/users.js
@@ -28,57 +28,7 @@ const saltRounds = 16;
 const router = Router();
 const geocoder = NodeGeocoder(geocoderConfig);
 
-router
-  .route("/leaderboard/local")
-  .get(async (req, res) => {
-    const userId = req.session.user && req.session.user._id;
-    const userCountryCode =
-      req.session.user &&
-      req.session.user.geocode &&
-      req.session.user.geocode.countryCode;
-    const userCity =
-      req.session.user &&
-      req.session.user.geocode &&
-      req.session.user.geocode.city;
 
-    let leaderboard;
-    try {
-      if (userId)
-        leaderboard = await topNthLocalUsersByHighScore(
-          100,
-          userCountryCode,
-          userCity
-        );
-      else
-        leaderboard = await topNthLocalUsersByHighScore(100, "US", "New York");
-    } catch (error) {
-      return res.status(500).send("Internal Server Error:", error);
-    }
-
-    return res.render("leaderboard", {
-      title: "Local Leaderboard",
-      leaderboard,
-    });
-  })
-  .post(async (req, res) => {
-    // reserved for AJAX
-  });
-
-router.route("/leaderboard/global").get(async (req, res) => {
-  const userId = req.session.user && req.session.user._id;
-
-  let leaderboard;
-  try {
-    leaderboard = await topNthGlobalUsersByHighScore(100);
-  } catch (error) {
-    return res.status(500).send("Internal Server Error:", error);
-  }
-
-  return res.render("leaderboard", {
-    title: "Global Leaderboard",
-    leaderboard,
-  });
-});
 
 router
   .route("/user")
@@ -157,157 +107,10 @@ router
 
     return res.redirect("/user/user_profile");
   });
-  router.
-  route('/gamestart')
-  .get(async (req,res)=>{
-      res.render('game_start',{title: 'Quiz'});
-      })
-  .post(async(req,res)=>{
-      if(req.session.user){
-          req.session.questions = await getQuestions4User({
-            numberOfOptions : 4,
-            numberOfQuestions : 50,
-            countryCode : req.session.user['geocode.countryCode'],
-            city : req.session.user['geocode.city'],
-          });
-      }else{
-          req.session.questions =await getQuestions4Guest({
-            numberOfOptions : 4,
-            numberOfQuestions : 50,
-          } );
-      }
-      req.session.index = 0;
-      req.session.score = 0;
-      req.session.timer = 60
-      res.redirect('/users/gameplay');
-  });
   
-router.
-  route('/gameplay')
+router
+  .route("/user/profile")
   .get(async (req, res) => {
-      if(!req.session.questions || !req.session.timer){
-        res.redirect('/users/gamestart');
-      }else{
-        let questions = req.session.questions;
-        let index = req.session.index;
-        let time = req.session.timer;
-        if(!req.session.score){
-            req.session.score=0;
-        }
-        if(questions.length<=index){
-            if(req.session.timer>0){
-                req.session.score = (req.session.score)*(60/(60-req.session.timer));
-            }
-            res.redirect('/users/gameresult');
-        }
-        let score = req.session.score;
-        return res.render('game_question', {title: 'Quiz',question : questions[index],index : index+1,score : score,time});
-      }
-      })
-  .post(async(req,res)=>{
-      let questions = req.session.questions;
-      let index = req.session.index;
-      const {options,timer} = req.body;
-      if(timer<=0){
-          if(options){
-              let correct = questions[index]['answer'];
-              if(questions[index]['options'][options] == correct){
-                  req.session.score = req.session.score + questions[index]['difficulty'];
-              }
-              req.session.timer = timer;
-          }
-          res.redirect('/users/gameresult');
-      }else{
-          let correct = questions[index]['answer'];
-          if(questions[index]['options'][options] == correct){
-              req.session.score = req.session.score + questions[index]['difficulty'];
-          }
-          req.session.index = index + 1;
-          req.session.timer = timer;
-          res.redirect('/users/gameplay')
-      }
-      });
-router.
-  route('/gameresult')
-  .get(async (req, res) => {
-    if(!req.session.questions){
-      res.redirect('/users/gamestart')
-    }else{
-        if(req.session.user){
-            let questions = req.session.questions;
-            let score = req.session.score;
-            let user =req.session.user;
-            for(let i=0;i<index;i++){
-                if(i<questions.length){
-                    let birdid = questions[i]['birdid'];
-                    await updatePlayerInfoById(user['_id'],{
-                      $pushLastQuestions: { birdid},
-                    });
-                }
-            }
-            delete req.session['questions'];
-            delete req.session['index'];
-            delete req.session['score'];
-            delete req.session['timer'];
-            res.render('game_end',{score});
-        }
-        else{
-            let score = req.session.score;
-            delete req.session['questions'];
-            delete req.session['index'];
-            delete req.session['score'];
-            delete req.session['timer'];
-            res.render('game_end',{score});
-        }
-      }
-      });
-      router
-      .route("/login")
-      .get(async (req, res) => {
-        const userId = req.session.user && req.session.user._id;
-        if (userId) return res.redirect("/user/user_profile");
-        return res.render("login", { title: "Login" });
-      })
-      .post(async (req, res) => {
-        const userId = req.session.user && req.session.user._id;
-        if (userId) return res.redirect("/user/user_profile");
-    
-        let { username, password } = req.body;
-        try {
-          username = checkUserName(username);
-          password = checkPassword(password);
-        } catch (error) {
-          return res
-            .status(400)
-            .render("login", { title: "Login", errors: [error] });
-        }
-    
-        let user;
-        try {
-          user = await getUserByUserName(username);
-        } catch (error) {
-          return res.status(500).send("Internal Server Error:", error);
-        }
-    
-        if (!user)
-          return res.status(400).render("login", {
-            title: "Login",
-            errors: ["Either username or password is incorrect!"],
-          });
-        if (!(await bcrypt.compare(password, user.hashed_password))) {
-          return res.status(400).render("login", {
-            title: "Login",
-            errors: ["Either username or password is incorrect!"],
-          });
-        }
-    
-        req.session.user = { _id: user._id, username: user.username };
-        return res.redirect("/user/user_profile")
-      });
-    
-    router
-      .route("/user/profile")
-      .get(async (req, res) => {
         const userId = req.session.user && req.session.user._id;
         if (!userId) return res.redirect("/login");
